seperate looping menu and function file
