+++ .gitignore
@@ -1,0 +1,2 @@
+Function/__pycache__/*
+Function/__pycache__/
